Restructuracion de proyecto correcta
