+++ src/infrastructure/database/user.repository.ts
@@ -1,15 +1,14 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaClient } from 'generated/prisma';
 import { PrismaService } from '../services/prisma.service';
 
-const prisma = new PrismaClient();
+
 
 @Injectable()
 export class UserRepository {
   constructor(private readonly prisma: PrismaService) {}
 
   async findByEmail(email: string) {
-    return prisma.user.findUnique({
+    return this.prisma.user.findUnique({
       where: { email },
     });
   }
