+++ src/modules/app.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { AppService } from '../infrastructure/services/app.service';
-import { AppController } from 'src/application/controllers/app.controller';
+import { AppController } from 'src/infrastructure/controllers/app.controller';
 import { AuthModule } from './auth.module';
 
 @Module({
