+++ src/modules/auth.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { AuthController } from '../application/controllers/auth.controller';
+import { AuthController } from '../infrastructure/controllers/auth.controller';
 import { AuthUseCase } from 'src/application/use-cases/auth.use-case';
 import { UserRepository } from 'src/infrastructure/database/user.repository';
 import { PrismaModule } from './prisma.module';
